requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beartype==0.18.5
billiard==4.2.0
bitrix24-rest==2.0.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
et-xmlfile==1.1.0
fast_bitrix24==1.7.1
frozenlist==1.4.1
icontract==2.6.6
idna==3.7
kombu==5.3.7
magic-filter==1.0.12
more-itertools==10.2.0
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.4
requests==2.31.0
six==1.16.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
